fix: :bug: Corrigidos valores long que estavam como inteiro

+++ TeleAppBotApi/TeleAppBot.Application/Mensagens/EnviarMensagem/EnviarMensagemCommand.cs
@@ -3,5 +3,5 @@ using TeleAppBot.Domain.ValueObjects;
 
 namespace TeleAppBot.Application.Mensagens.EnviarMensagem
 {
-    public record EnviarMensagemCommand(int IdMensagem, int IdChat, int IdContato, TipoMensagem Tipo, DateTime Data, InformacoesContatoCommand Contato, EnviarMensagemTextoCommand? MensagemTexto, EnviarMensagemMidiaCommand? MensagemMidia) : IRequest<Unit>;
+    public record EnviarMensagemCommand(int IdMensagem, long IdChat, long IdContato, TipoMensagem Tipo, DateTime Data, InformacoesContatoCommand Contato, EnviarMensagemTextoCommand? MensagemTexto, EnviarMensagemMidiaCommand? MensagemMidia) : IRequest<Unit>;
 }
