+++ TeleAppBotApi/TeleAppBot.Domain/DomainServices/IContatoDomainService.cs
@@ -2,8 +2,8 @@
 {
     public interface IContatoDomainService
     {
-        Task ValidarExistenciaDeContato(int idUsuario, bool eBot, string nome, string sobrenome, string usuario);
+        Task ValidarExistenciaDeContato(long idUsuario, bool eBot, string nome, string sobrenome, string usuario);
 
-        Task<bool> ValidarExistenciaDeContato(int idUsuario);
+        Task<bool> ValidarExistenciaDeContato(long idUsuario);
     }
 }
