+++ TeleAppBotApi/TeleAppBot.Domain/Entities/Conversas/Conversa.cs
@@ -12,12 +12,12 @@ namespace TeleAppBot.Domain.Entities.Conversas
 
         public IReadOnlyList<Mensagem> Mensagens => _mensagens.AsReadOnly();
 
-        public Conversa(Guid idContato, int idChat, IList<Mensagem> mensagens) : this(idContato, idChat)
+        public Conversa(Guid idContato, long idChat, IList<Mensagem> mensagens) : this(idContato, idChat)
         {
             _mensagens = mensagens;
         }
 
-        public Conversa(Guid idContato, int idChat)
+        public Conversa(Guid idContato, long idChat)
         {
             IdContato = idContato;
             IdChat = idChat;
