+++ TeleAppBotApi/TeleAppBot.Domain/Events/EnviarMensagemMidiaEvent.cs
@@ -3,5 +3,5 @@ using TeleAppBot.Domain.ValueObjects;
 
 namespace TeleAppBot.Domain.Events
 {
-    public record EnviarMensagemMidiaEvent(int IdMensagem, int IdChat, int IdContato, TipoMensagem Tipo, DateTime Data, string IdArquivo, string IdUnicoArquivo, int Tamanho) : Evento;
+    public record EnviarMensagemMidiaEvent(int IdMensagem, long IdChat, long IdContato, TipoMensagem Tipo, DateTime Data, string IdArquivo, string IdUnicoArquivo, int Tamanho) : Evento;
 }
