+++ TeleAppBotApi/TeleAppBot.Domain/Events/EnviarMensagemTextoEvent.cs
@@ -3,5 +3,5 @@ using TeleAppBot.Domain.ValueObjects;
 
 namespace TeleAppBot.Domain.Events
 {
-    public record EnviarMensagemTextoEvent(int IdMensagem, int IdChat, int IdContato, InformacoesUsuarioEvent InformacoesUsuario, TipoMensagem Tipo, DateTime Data, string Texto) : Evento;
+    public record EnviarMensagemTextoEvent(int IdMensagem, long IdChat, long IdContato, InformacoesUsuarioEvent InformacoesUsuario, TipoMensagem Tipo, DateTime Data, string Texto) : Evento;
 }
