+++ TeleAppBotApi/TeleAppBot.Domain/Repositories/IContatosRepository.cs
@@ -4,6 +4,6 @@ namespace TeleAppBot.Domain.Repositories
 {
     public interface IContatosRepository : IRepository<Contato>
     {
-        Task<Contato> ObterPorIdContato(int idContato);
+        Task<Contato> ObterPorIdContato(long idContato);
     }
 }
